requirements.txt
absl-py==0.8.1
alabaster==0.7.12
alembic==1.0.11
asn1crypto==0.24.0
astor==0.8.0
async-generator==1.10
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blinker==1.4
blis==0.4.1
bokeh==1.3.4
branca==0.3.1
bs4==0.0.1
certifi==2019.9.11
certipy==0.1.3
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
conda==4.7.10
conda-package-handling==1.4.0
coverage==4.5.3
coveralls==1.8.2
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
cytoolz==0.10.1
dask==2.2.0
datascience==0.15.3
decorator==4.4.0
defusedxml==0.5.0
dill==0.3.0
distributed==2.2.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
fastcache==1.1.0
folium==0.10.0
fsspec==0.4.1
fuzzywuzzy==0.17.0
gast==0.3.2
gmpy2==2.1.0b1
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.16.1
h5py==2.9.0
heapdict==1.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-tensorboard==0.1.10
jupyterhub==1.0.0
jupyterlab==1.0.4
jupyterlab-server==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Mako==1.0.14
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
murmurhash==1.0.0
nbcluster-status==0.0.0
nbconvert==5.6.1
nbformat==4.4.0
nbgitpuller==0.7.2
nbgrader==0.6.0
nbresuse==0.2.0
networkx==2.3
nltk==3.4.4
notebook==6.0.0
numba==0.45.1
numexpr==2.6.9
numpy==1.16.4
oauthlib==3.0.1
okpy==1.14.18
olefile==0.46
opencv-contrib-python-headless==3.4.5.20
packaging==19.0
pamela==1.0.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
pluggy==0.13.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycosat==0.6.3
pycparser==2.19
pycurl==7.43.0.2
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyrsistent==0.15.4
PySocks==1.7.0
pytest==5.2.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.5
requests==2.12.4
ruamel-yaml==0.15.71
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.2
spacy==2.2.2
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.6
srsly==0.2.0
statsmodels==0.10.1
sympy==1.4
tblib==1.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.0
toolz==0.10.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
vincent==0.4.4
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zict==1.0.0
zipp==0.6.0
